init `contactsData` constant
